refactor : 분리되어 있던 상품정보 타입을 합치기
